requirements.txt
alembic==1.16.1
annotated-types==0.7.0
anyio==4.9.0
click==8.2.1
fastapi==0.115.12
h11==0.16.0
idna==3.10
python-jose[cryptography]==3.3.0
Mako==1.3.10
MarkupSafe==3.0.2
passlib==1.7.4
pydantic==2.11.5
pydantic-settings==2.9.1
pydantic_core==2.33.2
python-dotenv==1.1.0
sniffio==1.3.1
SQLAlchemy==2.0.41
starlette==0.46.2
typing-inspection==0.4.1
typing_extensions==4.14.0
uvicorn==0.34.3
pydantic[email]
asyncpg==0.29.0
psycopg2-binary==2.9.9
python-multipart==0.0.11
celery==5.3.4
redis==5.0.1
websockets
jinja2
openai>=1.0.0
aiofiles
aiohttp